backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bcrypt==4.1.3
black==25.12.0
boto3==1.42.21
botocore==1.42.21
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
emergentintegrations==0.1.0
fastapi==0.110.1
fastuuid==0.14.0
filelock==3.20.2
flake8==7.3.0
frozenlist==1.8.0
fsspec==2025.12.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.57.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
http_ece==1.2.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface_hub==1.2.4
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
isort==7.0.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
jq==1.10.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
librt==0.7.7
litellm==1.80.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
multidict==6.7.0
mypy==1.19.1
mypy_extensions==1.1.0
numpy==2.4.0
oauthlib==3.3.1
openai==1.99.9
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
py-vapid==1.9.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pyparsing==3.3.1
pypdf==6.6.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
pytokens==0.3.0
pytz==2025.2
pywebpush==2.2.0
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
resend==2.19.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
s3transfer==0.16.0
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
stripe==14.1.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.1
typer==0.21.0
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.25.0
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
